Review fixes

+++ test/unit/test_utils.h
@@ -4,11 +4,11 @@
 #include <cstdint>
 
 inline int32_t make_timestamp(int year_value,
-                              unsigned month_value,
-                              unsigned day_value,
-                              unsigned hour_value,
-                              unsigned minute_value,
-                              unsigned second_value)
+                              uint8_t month_value,
+                              uint8_t day_value,
+                              uint8_t hour_value,
+                              uint8_t minute_value,
+                              uint8_t second_value)
 {
     using namespace std::chrono;
     const sys_days date = year{year_value} / month{month_value} / day{day_value};
